Fix unit normalization, add stop_event to monitor, and harden validation
- api_server: always return unit="C" — temperature_c/temperature_f are already
  normalized, so the raw sensor unit was misleading (bug: live F-unit sensors
  would return unit="F" while temperature_c held Celsius)
- temperature_sensor: validate sensor_id non-empty in TemperatureReading
- temperature_sensor: validate get_all_sensors() response structure
- temperature_sensor: monitor() accepts stop_event for cooperative cancellation;
  replaces time.sleep with stop_event.wait so it wakes immediately on signal
- config: POLL_INTERVAL cast to float to match monitor()'s interval: float type
- ui: _clear_log only clears scrollback — stats (min/max/count) track the session
- ui: timestamps labeled UTC throughout

+++ config.py
@@ -5,7 +5,7 @@ DEMO_MODE = os.environ.get("DEMO_MODE", "true").lower() == "true"
 API_BASE_URL = os.environ.get("API_BASE_URL", "https://api.example.com/v1")
 API_KEY = os.environ.get("API_KEY", "")
 SENSOR_ID = os.environ.get("SENSOR_ID", "sensor_001")
-POLL_INTERVAL = int(os.environ.get("POLL_INTERVAL", "2"))
+POLL_INTERVAL = float(os.environ.get("POLL_INTERVAL", "2"))
 POLL_CYCLES = int(os.environ.get("POLL_CYCLES", "8"))
 
 if not (1 <= POLL_INTERVAL <= 3600):
